+++ src/main/java/com/samb1232/urfu_java_bot/BotApplication.java
@@ -4,8 +4,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class UrfuJavaBotApplication {
+public class BotApplication {
 	public static void main(String[] args) {
-		SpringApplication.run(UrfuJavaBotApplication.class, args);
+		SpringApplication.run(BotApplication.class, args);
 	}
 }
